server/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
decouple==0.0.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
ruff==0.1.2
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
uvicorn==0.15.0
urllib3==2.0.6
whitenoise==6.6.0
